+++ library/functions/enqueue-scripts-styles.php
@@ -19,9 +19,6 @@ function sr_enqueue_scripts_styles()
 	// Adding Foundation scripts file in the footer
 	wp_enqueue_script( 'foundation', get_template_directory_uri() . '/assets/js/foundation.min.js', array( 'jquery', 'what-input' ), '6.2.3', true );
 	
-	// Adding fontfaceboserver scripts file in the header
-	wp_enqueue_script( 'fontfaceobserver', get_template_directory_uri() . '/library/vendor/fontfaceobserver/fontfaceobserver.js', array( ), '1.7.1', false );	// Fontfaceobserver from https://github.com/bramstein/fontfaceobserver
-	wp_enqueue_script( 'sr-fontface', get_template_directory_uri() . '/assets/js/fontface.min.js', array( 'fontfaceobserver' ), '1.0', false );					// Will be overwritten in child theme
 	
 	// enqueue comment js
 	if ( is_singular() && comments_open() && get_option( 'thread_comments' ) ) {